# backend/swagger.yaml
openapi: 3.0.0
info:
  title: API Baclemd - Proxy per microservices
  description: API aggregata per i microservizi di un e-commerce artigianale
  version: 1.0.0
servers:
  - url: https://localhost:3000
    description: API Gateway principale
...
security:
  - cookieAuth: []
paths:
...
  /user:
    post:
      summary: Registrazione di un nuovo utente
      tags:
        - Users
      description: Permette la registrazione di un nuovo utente con ruolo 'customer' o 'artisan'.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - surname
                - email
                - password
                - role
              properties:
                name:
                  type: string
                  example: Mario
                surname:
                  type: string
                  example: Rossi
                email:
                  type: string
                  format: email
                  example: mario.rossi@example.com
                password:
                  type: string
                  format: password
                  example: securePassword123!
                role:
                  type: string
                  enum: [customer, artisan]
                  description: Ruolo dell'utente (cliente o artigiano).
                  example: customer
      responses:
        '200':
          description: Utente registrato con successo.
          headers:
            Set-Cookie:
              schema:
                type: string
                example: jwt=your_jwt_token; Path=/; HttpOnly; Secure; SameSite=Lax
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    example: 1
                  email:
                    type: string
                    format: email
                    example: mario.rossi@example.com
                  name:
                    type: string
                    example: Mario
                  surname:
                    type: string
                    example: Rossi
                  id_role:
                    type: integer
                    example: 1
                  roleName:
                    type: string
                    example: customer
        '400':
          $ref: '#/components/responses/BadRequest'
        '512':
          description: Violazione di unicità (email già registrata).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 512
                message: Email already registered.
        '516':
          description: Formato password non valido.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 516
                message: Invalid password format.
        '517':
          description: Formato email non valido.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 517
                message: Invalid email format.
        '518':
          description: Errore durante la generazione dello slug utente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 518
                message: Error generating user slug.
        '522':
          description: Ruolo utente non valido.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 522
                message: Invalid user role.
        '500':
          $ref: '#/components/responses/InternalServerError'
    get:
      summary: Ottiene il profilo dell'utente autenticato
      tags:
        - Users
      description: Recupera le informazioni del profilo dell'utente autenticato tramite JWT.
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Informazioni del profilo recuperate con successo.
          content:
            application/json:
              schema:
                type: object
                properties:
                  name:
                    type: string
                    example: Mario
                  surname:
                    type: string
                    example: Rossi
                  role:
                    type: string
                    example: customer
                  bio:
                    type: string
                    nullable: true
                    example: Sono un artigiano specializzato in ceramica.
                  url_profile_picture:
                    type: string
                    format: url
                    nullable: true
                    example: https://localhost:3000/images/profile_pic_id
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '515':
          description: Utente non trovato (per l'ID nel JWT).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 515
                message: User not found.
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      summary: Aggiorna il profilo utente autenticato
      tags:
        - Users
      description: Aggiorna le informazioni del profilo dell'utente autenticato.
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  description: Nuovo indirizzo email (verrà convalidato).
                  example: nuovo.email@example.com
                name:
                  type: string
                  example: Luigi
                surname:
                  type: string
                  example: Verdi
                bio:
                  type: string
                  nullable: true
                  example: Nuova descrizione della biografia.
                password:
                  type: string
                  format: password
                  description: Nuova password (verrà hashata).
                id_profile_picture:
                  type: string
                  format: uuid
                  nullable: true
                  description: ID della nuova immagine del profilo.
                  example: 123e4567-e89b-12d3-a456-426614174000
              minProperties: 1 # Almeno un campo deve essere presente per l'aggiornamento
      responses:
        '200':
          description: Profilo aggiornato con successo.
          content:
            application/json:
              schema:
                type: object
                properties:
                  password_changed:
                    type: boolean
                    description: Indica se la password è stata modificata.
                    example: true
                  email:
                    type: string
                    format: email
                    example: nuovo.email@example.com
                  name:
                    type: string
                    example: Luigi
                  surname:
                    type: string
                    example: Verdi
                  bio:
                    type: string
                    example: Nuova descrizione della biografia.
                  id_profile_picture:
                    type: string
                    format: uuid
                    example: 123e4567-e89b-12d3-a456-426614174000
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '512':
          description: L'email fornita è già in uso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 512
                message: Email already registered.
        '516':
          description: Formato password non valido.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 516
                message: Invalid password format.
        '517':
          description: Formato email non valido.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 517
                message: Invalid email format.
        '521':
          description: ID immagine del profilo non valido o non esistente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 521
                message: Invalid image ID.
        '500':
          $ref: '#/components/responses/InternalServerError'
  /login:
    post:
      summary: Login utente
      tags:
        - Users
      description: Autentica un utente tramite email e password e imposta un cookie JWT.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: mario.rossi@example.com
                password:
                  type: string
                  format: password
                  example: securePassword123!
      responses:
        '200':
          description: Login riuscito. Il cookie JWT è impostato.
          headers:
            Set-Cookie:
              schema:
                type: string
                example: jwt=your_jwt_token; Path=/; HttpOnly; Secure; SameSite=Lax
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    example: 1
                  email:
                    type: string
                    format: email
                    example: mario.rossi@example.com
                  name:
                    type: string
                    example: Mario
                  surname:
                    type: string
                    example: Rossi
                  id_role:
                    type: integer
                    example: 1
                  roleName:
                    type: string
                    example: customer
        '400':
          $ref: '#/components/responses/BadRequest'
        '513':
          description: Utente non trovato (email non registrata).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 513
                message: User not found.
        '514':
          description: Password non corretta.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 514
                message: Incorrect password.
        '500':
          $ref: '#/components/responses/InternalServerError'
  /logout:
    post:
      summary: Logout utente
      tags:
        - Users
      description: Effettua il logout dell'utente invalidando il cookie JWT.
      responses:
        '200':
          description: Logout riuscito. Il cookie JWT è invalidato.
          headers:
            Set-Cookie:
              schema:
                type: string
                example: jwt=; Path=/; Expires=Thu, 01 Jan 1970 00:00:00 GMT; HttpOnly; Secure; SameSite=Lax
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /isLoggedIn:
    get:
      summary: Verifica lo stato di login dell'utente
      tags:
        - Users
      description: Controlla se l'utente è attualmente loggato, verificando la presenza del cookie JWT.
      responses:
        '200':
          description: Stato di login restituito.
          content:
            application/json:
              schema:
                type: object
                properties:
                  loggedIn:
                    type: boolean
                    example: true
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: jwt
      description: Autenticazione tramite cookie HTTP-only contenente il JWT.
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          description: Codice di errore specifico del backend.
          example: 500
        message:
          type: string
          description: Descrizione leggibile dell'errore.
          example: Internal Server Error.
...
  responses:
    BadRequest:
      description: Richiesta non valida. Campi obbligatori mancanti o formattati male.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: 400
            message: Bad Request.
    UnauthorizedError:
      description: Autenticazione richiesta o non valida (token JWT mancante/scaduto/non valido).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: 401
            message: Unauthorized.
...
    InternalServerError:
      description: Errore interno del server.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: 500
            message: Internal Server Error.